Reorganizate directories. #2
