+++ src/controllers/auth/emailConfirm.controller.ts
@@ -1,14 +1,14 @@
 import dotenv from 'dotenv';
-import * as AuthUtil from '../utils/AuthUtil.utils';
-import { AuthMessages } from '../utils/AuthMessage.util';
-import { UserRepository } from '../repository/implementations/user.implementation';
+import { AuthUtil} from '../../utils/auth/AuthUtil.utils';
+import { RedisUtil } from '../../utils/redis/RedisUtils.util';
+import { AuthMessages } from '../../utils/messages/AuthMessage.enum';
 
 dotenv.config();
 export async function confirmUserEmail(ctx: any): Promise<any> {
     const token = ctx.params.token;
     try {
         if (typeof(token) === "string") {
-            const user = await AuthUtil.RedisExistKey(process.env.EMAIL_VERIFICATION_TOKEN+token);
+            const user = await RedisUtil.RedisExistKey(process.env.EMAIL_VERIFICATION_TOKEN+token);
             if (user) {
                 return AuthUtil.handleSuccessLogout(AuthMessages.EmailIsConfirmed, ctx);
             } else {
