+++ src/controllers/auth/logout.controller.ts
@@ -1,6 +1,7 @@
 
-import * as AuthUtil from '../utils/AuthUtil.utils';
-import { AuthMessages } from '../utils/AuthMessage.util';
+import { AuthUtil } from '../../utils/auth/AuthUtil.utils';
+import { RedisUtil } from '../../utils/redis/RedisUtils.util';
+import { AuthMessages } from '../../utils/messages/AuthMessage.enum';
 
 export async function logoutUser(ctx: any): Promise<any> 
 {
@@ -8,13 +9,13 @@ export async function logoutUser(ctx: any): Promise<any>
     try {
        if(typeof(token) === "string")
        {
-         const deleteAuth = await AuthUtil.RedisDeleteKey(token); 
+         const deleteAuth = await RedisUtil.RedisDeleteKey(`${process.env.AUTH_REDIS_REC}`+token); 
         return deleteAuth === true 
                ? AuthUtil.handleSuccessLogout(AuthMessages.LogoutSuccess,ctx)
                : AuthUtil.handleAppValidation(AuthMessages.LogoutFailed,ctx);
        } 
     } catch (error:any) {
-        console.debug(error);
+         console.debug(error);
         return AuthUtil.handleInternalServerError(AuthMessages.InternalServerError,ctx);
     } 
 }
