+++ src/controllers/auth/register.controller.ts
@@ -1,23 +1,22 @@
 'use strict';
 
 import dotenv from 'dotenv';
-import {validateNewUser} from './validations/controller.validations';
-import { UserRepository } from '../repository/implementations/user.implementation';
-import * as AuthUtil from '../utils/AuthUtil.utils';
-import { AuthMessages } from '../utils/AuthMessage.util';
+import { AuthValidator} from '../validations/auth/auth.validations';
+import { UserRepository } from '../../repository/implementations/user.implementation';
+import { AuthUtil } from '../../utils/auth/AuthUtil.utils';
+import { AuthMessages } from '../../utils/messages/AuthMessage.enum';
 
 dotenv.config();
-
 const user:UserRepository = new UserRepository();
 
 export async function createNewUser(ctx: any): Promise<any> {
     try{
-        const registrationFormValidation = await validateNewUser(ctx.request.body);
+        const registrationFormValidation = await AuthValidator.validateNewUser(ctx.request.body);
         if (registrationFormValidation){
            const encryptPassword = await AuthUtil.encryptPassword(ctx.request.body.password);
            const result = await user.createUser(ctx.request.body,encryptPassword,"user");
             if(result === true){
-                  const { token, redisKey, link } = await prepareEmailVerification(ctx.request.body);
+                  const { token, redisKey, link } = await AuthUtil.prepareEmailVerification(ctx.request.body);
                 return AuthUtil.handleSuccessfulRegister(AuthMessages.UserCreatedSuccessfully,ctx)
             }else{
                return  AuthUtil.handleAppValidation(AuthMessages.UserAlreadyExists,ctx);
@@ -29,11 +28,4 @@ export async function createNewUser(ctx: any): Promise<any> {
           console.log(error); 
         return AuthUtil.handleInternalServerError(AuthMessages.InternalServerError,ctx);
     }    
-}
-
-async function prepareEmailVerification(userData: any): Promise<{ token: string; redisKey: string; link: string }> {
-    const token = await AuthUtil.generateTokenForEmailVerification(userData);
-        const redisKey = await AuthUtil.createRedisRecordForConfirmationEmail(userData, token);
-        const link = AuthUtil.generateLinkForEmailVerification(token);
-    return { token, redisKey, link };
 }
