+++ src/repository/implementations/role.implementation.ts
@@ -1,10 +1,10 @@
-import Roles from "../../models/roles.models";
-import {IRoleRepository} from "../interfaces/IRoleRepository.interface"
+import { Roles } from "../../models/roles.models";
+import { IRoleRepository } from "../interfaces/IRoleRepository.interface"
 
-export class RoleRepository implements IRoleRepository
-{
+ export class RoleRepository implements IRoleRepository
+ {
     constructor (){}
     async findRole(name:string): Promise<any | null> {
        return Roles.findOne({where: {Name: name}});
     }
-}
+ }
